// instaclone-web/src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
const defaultOptions = {} as const;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};

export type CreateAccountResult = {
  __typename?: 'CreateAccountResult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};

export type CreateCommentResult = {
  __typename?: 'CreateCommentResult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};

export type EditPhotoResult = {
  __typename?: 'EditPhotoResult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};

export type EditProfileResult = {
  __typename?: 'EditProfileResult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};

export type FollowUserResult = {
  __typename?: 'FollowUserResult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};
// ...
export type LoginResult = {
  __typename?: 'LoginResult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
  token?: Maybe<Scalars['String']>;
};

export type Message = {
  __typename?: 'Message';
  createdAt: Scalars['String'];
  id: Scalars['Int'];
  payload: Scalars['String'];
  read: Scalars['Boolean'];
  room: Room;
  updatedAt: Scalars['String'];
  user: User;
};

export type Mutation = {
  __typename?: 'Mutation';
  createAccount: CreateAccountResult;
  createComment?: Maybe<CreateCommentResult>;
  deleteComment: DeleteCommentResoult;
  deletePhoto: DeletePhotoResult;
  editComment: MutationResponse;
  editPhoto: EditPhotoResult;
  editProfile: EditProfileResult;
  followUser?: Maybe<FollowUserResult>;
  login: LoginResult;
  readMessage: MutationResponse;
  sendMessage: MutationResponse;
  toggleLike?: Maybe<ToggleLikeResult>;
  unfollowUser?: Maybe<UnfollowUserResult>;
  uploadPhoto?: Maybe<Photo>;
};
// ...
export type MutationResponse = {
  __typename?: 'MutationResponse';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};

export type Photo = {
  __typename?: 'Photo';
  caption?: Maybe<Scalars['String']>;
  comments: Scalars['Int'];
  createdAt: Scalars['String'];
  file: Scalars['String'];
  hashtags?: Maybe<Array<Maybe<Hashtag>>>;
  id: Scalars['Int'];
  isMine: Scalars['Boolean'];
  likes: Scalars['Int'];
  updatedAt: Scalars['String'];
  user: User;
};
// ...
export type UnfollowUserResult = {
  __typename?: 'UnfollowUserResult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};
// ...
export type DeleteCommentResoult = {
  __typename?: 'deleteCommentResoult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};

export type DeletePhotoResult = {
  __typename?: 'deletePhotoResult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};
// ...
export type ToggleLikeResult = {
  __typename?: 'toggleLikeResult';
  error?: Maybe<Scalars['String']>;
  ok: Scalars['Boolean'];
};

export type LoginMutationVariables = Exact<{
  userName: Scalars['String'];
  password: Scalars['String'];
}>;


export type LoginMutation = { __typename?: 'Mutation', login: { __typename?: 'LoginResult', ok: boolean, token?: string | null, error?: string | null } };

export type CreateAccountMutationVariables = Exact<{
  userName: Scalars['String'];
  lastName?: InputMaybe<Scalars['String']>;
  firstName: Scalars['String'];
  email: Scalars['String'];
  password: Scalars['String'];
}>;


export type CreateAccountMutation = { __typename?: 'Mutation', createAccount: { __typename?: 'CreateAccountResult', ok: boolean, error?: string | null } };